# app/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payment Validator API - Documentación del Sistema",
    "description": "API de conciliación de pagos y manejo de Webhooks para la pasarela de procesamiento Payment Validator.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "paths": {
    "/api/transacciones/validar": {
      "post": {
        "summary": "Validar y conciliar una transacción de Binance Pay",
        "description": "Cruza los datos ingresados por el cajero con los registros pendientes inyectados por el Worker. La búsqueda se realiza estrictamente por monto, moneda y el sufijo del ID de Binance (mínimo 6 dígitos). Si se encuentra, se vincula al número de venta/factura.",
        "tags": ["Transacciones"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "monto",
                  "moneda",
                  "binance_id_sufijo",
                  "fecha_cajero"
                ],
                "properties": {
                  "id_orden": {
                    "type": "string",
                    "nullable": true,
                    "description": "Número de venta o factura interna a la cual pertenece la transacción (Opcional).",
                    "example": "VENT-1024"
                  },
                  "monto": {
                    "type": "number",
                    "format": "float",
                    "description": "Monto exacto del pago recibido.",
                    "example": 25.50
                  },
                  "moneda": {
                    "type": "string",
                    "description": "Criptomoneda utilizada en el pago.",
                    "enum": ["USDT", "BTC", "ETH"],
                    "example": "USDT"
                  },
                  "binance_id_sufijo": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 12,
                    "description": "Sufijo del ID completo de la transacción de Binance: mínimo los últimos 6 dígitos que tipea el cajero, hasta 12 cuando el OCR logra leer más (reduce falsos positivos entre dos IDs que coinciden en los últimos 6 dígitos).",
                    "example": "4310701056"
                  },
                  "fecha_cajero": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Fecha y hora local en formato ISO enviada desde el cliente del cajero.",
                    "example": "2026-03-31T20:15:00.000Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pago verificado y conciliado con éxito.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Pago verificado con éxito"
                    },
                    "notificacionTelegram": {
                      "type": "boolean",
                      "description": "false si el pago quedó registrado pero el aviso al grupo de Telegram no pudo enviarse.",
                      "example": true
                    },
                    "transaccion": {
                      "type": "object",
                      "description": "Objeto completo de la transacción actualizada en la base de datos."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios o no se encontró ninguna coincidencia pendiente en Binance."
          },
          "409": {
            "description": "Conflicto de Idempotencia: El pago ya fue reclamado y verificado previamente."
          },
          "500": {
            "description": "Error interno del servidor."
          },
          "503": {
            "description": "Sistema suspendido, o fallo temporal de conexión con la base de datos: se puede reintentar."
          }
        }
      }
    },
...
  }
}